Conditions now can be used by other plugins
To make other plugins to be able extend you need 

1) Define class that is marked [ConditionModule]

2) Add method in that class is marked by [Condition("name", helpText: "tooltip")]

3) Method should return bool, and accept MyCubeGrid. It may also accept 1 extra param (string, double, int ...)
