+++ Essentials/Conditions/ConditionModule.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace Essentials
+{
+    [AttributeUsage(AttributeTargets.Class, AllowMultiple = false)]
+    public sealed class ConditionModule : Attribute
+    {
+        
+    }
+}
